fix(iso-date-adapter): use the locale of the date adapter (#152)
instead of the global day.js locale.

day.js differentiates between global and local locales. setLocale() in the date adapter doesn't
actually change the global day.js locale.
It did however rely on the global one in normalizeFormat().
If the global day.js locale was different to the one in the adapter, this would lead to
unexpected results.

This fix changes the behavior to use the adapter locale instead of the global day.js one.
If the formats of the current local are not available in day.js, it will
fallback to the global locale.