feat(number-stepper): add readonly state and rename former readonly to inputFieldReadonly (#1411)
BREAKING CHANGE:
the `readonly` input got renamed to `inputFieldReadonly`. the `readonly` property before this change only set the input field to readonly but left the number stepper operational via the buttons. this was not properly named and was changed now to make room for a proper readonly state that got added.